{{cookiecutter.project_slug}}/requirements/base.txt
# Versions are left out on purpose in order for you to use the latest ones.
# After the first "pip install -r", just run "pip freeze" and add the version
# to each package in each requirements/*.txt.

Django==1.7.11
django-environ==0.4.0
django-extensions==1.3.10
django-braces==1.4.0
django-model-utils==2.3.1
Fabric==1.10.2

-e git://github.com/huogerac/django-onedeploy.git#egg=django-onedeploy
